# backend/src/main/java/io/github/yvlar/coroute/dto/request/ConnexionRequest.java
package io.github.yvlar.coroute.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record ConnexionRequest(@Email @NotBlank String email, @NotBlank String motDePasse) {}