# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CoRoute API
  version: 1.0.0
  description: >
    API REST de covoiturage CoRoute. Décrit les endpoints réellement disponibles
    dans le backend JAX-RS/Jersey.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Accès direct au backend
  - url: /api
    description: Via le proxy Nginx/Vite
...
paths:
  /utilisateurs/inscription:
    post:
      tags: [Utilisateurs]
      summary: Créer un compte
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InscriptionRequest'
      responses:
        '201':
          description: Compte créé (sans corps)
        '400':
          description: Corps invalide
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Un compte existe déjà pour ce courriel
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /utilisateurs/connexion:
    post:
      tags: [Utilisateurs]
      summary: Authentifier un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnexionRequest'
      responses:
        '200':
          description: Jeton JWT et utilisateur
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '400':
          description: Corps invalide
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /trajets/match:
    get:
      tags: [Matching]
      summary: Trajets compatibles avec un besoin
      parameters:
        - { in: query, name: depart, schema: { type: string } }
        - { in: query, name: destination, schema: { type: string } }
        - in: query
          name: jours
          description: Répétable, ex. ?jours=LUNDI&jours=MARDI
          schema:
            type: array
            items: { $ref: '#/components/schemas/JourSemaine' }
      responses:
        '200':
          description: Liste de correspondances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MatchingResponse' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    JourSemaine:
      type: string
      enum: [LUNDI, MARDI, MERCREDI, JEUDI, VENDREDI, SAMEDI, DIMANCHE]
    TrajetType:
      type: string
      enum: [PONCTUEL, REGULIER]
    InscriptionRequest:
      type: object
      required: [nom, email, motDePasse]
      properties:
        nom: { type: string }
        email: { type: string, format: email }
        motDePasse: { type: string, minLength: 6 }
    ConnexionRequest:
      type: object
      required: [email, motDePasse]
      properties:
        email: { type: string, format: email }
        motDePasse: { type: string }
    UtilisateurResponse:
      type: object
      properties:
        id: { type: string }
        nom: { type: string }
        email: { type: string, format: email }
    TokenResponse:
      type: object
      properties:
        token: { type: string }
        utilisateur: { $ref: '#/components/schemas/UtilisateurResponse' }
...
    MatchingResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        depart: { type: string }
        destination: { type: string }
        heure: { type: string }
        prixParPassager: { type: number, format: double }
        placesRestantes: { type: integer }
        type: { $ref: '#/components/schemas/TrajetType' }
        joursCompatibles:
          type: array
          items: { $ref: '#/components/schemas/JourSemaine' }
        scoreCompatibilite: { type: integer }
        conducteurId: { type: string }
        date: { type: string, format: date, nullable: true }
        dateDebut: { type: string, format: date, nullable: true }
        dateFin: { type: string, format: date, nullable: true }
...
    ErrorResponse:
      type: object
      properties:
        message: { type: string }